requirements.txt
# UNet Polygon Coloring Project Requirements
# ==========================================

# PyTorch (will install CPU version by default, see setup.sh for GPU versions)
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Core dependencies
numpy>=1.24.0
Pillow>=10.0.0
wandb>=0.15.0
scikit-image>=0.20.0
tqdm>=4.65.0
matplotlib>=3.7.0

# Optional but recommended
tensorboard>=2.14.0
opencv-python>=4.8.0